fix : Id 타입 변경

+++ src/main/java/com/example/demo/volunteer/VolunteeringRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VolunteeringRepository extends JpaRepository<Volunteering, Integer> {
+public interface VolunteeringRepository extends JpaRepository<Volunteering, Long> {
     @Override
     List<Volunteering> findAll();
